reth_buffer/pyproject.toml
[tool.poetry]
name = "reth_buffer"
version = "0.1.0"
description = ""
authors = ["reth <sosp20212021@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.6 <3.10"
numba = "^0.53.0"
numpy = "^1.19.2"
psutil = "^5.7.2"
portpicker = "^1.3.1"
pyzmq = "^20.0.0"
netifaces = "^0.10.9"
lz4 = "^3.1.3"
loguru = "^0.5.3"
msgpack = "^1.0.2"
lmdb = "^1.1.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
